Fixed-width types where a width matters: 32-bit IOCTL codes, error numbers, process ID, atomics and symbol values, uint32_t and uint8_t fields in CDtapiLite.h; counters stay int

+++ Examples/Common/ExampleCommon.c
@@ -137,7 +137,7 @@ bool ExampleInt64(int Argc, char** Argv, const char* Name, int64_t* Value)
 {
     const char* Text = ExampleValue(Argc, Argv, Name);
     char* End = NULL;
-    long long Parsed;
+    int64_t Parsed;
 
     if (Text == NULL)
         return true;
